Fix more compile errors.

+++ Source/SampleConvolutionRenderer.cpp
@@ -106,7 +106,7 @@ void CSampleConvolutionRenderer::Execute( const SRenderContext& renderContext, c
 
 
     SD3D12DescriptorHandle SRVs[] = { scene.m_SamplePositionTexture->GetSRV(), scene.m_SampleValueTexture->GetSRV() };
-    D3D12_GPU_DESCRIPTOR_HANDLE descriptorTable = s_DescriptorTableLayout.AllocateAndCopyToGPUDescriptorHeap( SRVs, ARRAY_LENGTH( SRVs ), &scene.m_FilmTexture->GetUAV(), 1 );
+    D3D12_GPU_DESCRIPTOR_HANDLE descriptorTable = s_DescriptorTableLayout.AllocateAndCopyToGPUDescriptorHeap( SRVs, (uint32_t)ARRAY_LENGTH( SRVs ), &scene.m_FilmTexture->GetUAV(), 1 );
     commandList->SetComputeRootDescriptorTable( 1, descriptorTable );
 
     commandList->SetPipelineState( m_PSO.get() );
